requirements.txt
spotipy
numpy
pandas
python-dotenv
